% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_data.R
\docType{data}
\name{adae_data}
\alias{adae_data}
\title{Adverse event data set included in AdEPro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 683 rows and 16 columns.
}
\usage{
adae_data
}
\description{
Adverse event data set included in AdEPro
}
\keyword{datasets}
