% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adepro-package.R
\docType{package}
\name{adepro-package}
\alias{adepro}
\alias{adepro-package}
\title{adepro: A 'shiny' Application for the (Audio-)Visualization of Adverse Event Profiles}
\description{
Contains a 'shiny' application called AdEPro (Animation of Adverse Event Profiles) which (audio-)visualizes adverse events occurring in clinical trials. As this data is usually considered sensitive, this tool is provided as a stand-alone application that can be launched from any local machine on which the data is stored.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bayer-Group/BIC-AdEPro}
  \item Report bugs at \url{https://github.com/Bayer-Group/BIC-AdEPro/issues}
}

}
\author{
\strong{Maintainer}: Nicole Rethemeier \email{nicole.rethemeier@bayer.com}

Authors:
\itemize{
  \item Christoph Tasto
  \item Steffen Jeske
  \item Bodo Kirsch
}

}
\keyword{internal}
