% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_adeproLogo.R
\name{adeproLogo}
\alias{adeproLogo}
\title{adeproLogo  Function that creates a div container for the AdEPro Logo}
\usage{
adeproLogo(height = 230, width = 160, align = "right")
}
\arguments{
\item{height}{height of the Logo (numeric)}

\item{width}{width of the Logo (numeric)}

\item{align}{Alignment of the Logo ("center"/"left"/"right)}
}
\description{
creates a div container with AdEPro Logo for the shiny app adepro
}
\keyword{internal}
