% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_data.R
\docType{data}
\name{adsl_data}
\alias{adsl_data}
\title{Subject level data set included in AdEPro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 9 columns.
}
\usage{
adsl_data
}
\description{
Subject level data set included in AdEPro
}
\keyword{datasets}
