% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_count.R
\name{ae_count}
\alias{ae_count}
\title{ae_count - provides adverse event count dataset}
\usage{
ae_count(ae_data, patient)
}
\arguments{
\item{ae_data}{adverse event dataset}

\item{patients}{patient dataset}
}
\description{
Provides data set with one adverse event by day and by treatment groups with variable 'freq', giving a
categorization (integer between 0 and 4) of the frequency of the adverse event on a specific day by treatment
group.
}
\keyword{internal}
