% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_and_prepare_patient_data.R
\name{as_numeric_date}
\alias{as_numeric_date}
\title{Obtains date as number}
\usage{
as_numeric_date(variable)
}
\arguments{
\item{variable}{A vector, the variable}
}
\description{
Obtains date as number
}
