% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{bar_chart - creates barchart of AEs for all patient}
\usage{
bar_chart(
  ae_data = ae,
  patients = patient,
  day = day_slider,
  variables = vars,
  day_max = day_mx,
  count_max = count_mx,
  treatments = treatments,
  cex.n = 2,
  color_theme = TRUE
)
}
\arguments{
\item{ae_data}{Adverse event dataset}

\item{patients}{Patient dataset}

\item{day}{The study day of interest}

\item{variables}{Vector of Adverse events}

\item{day_max}{Maximum Day}

\item{count_max}{Maximum Counts}

\item{cex.n}{Font size of the text in the bars}

\item{color_theme}{logical value for dark/light theme}
}
\description{
Creates bar charts daily and total for all patients separately according to treatment (R Package required: 'dplyr')
}
\keyword{ae's}
\keyword{internal}
\keyword{non}
\keyword{refactor'}
\keyword{remove}
\keyword{to}
\keyword{used}
