% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_legend.R
\name{barchart_legend}
\alias{barchart_legend}
\title{barchart_legend - Creates Legend for adverse events barchart view in AdEPro}
\usage{
barchart_legend(
  tmp,
  aes,
  colors = c("#e43157", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33", "#a65628",
    "#f781bf", "#21d4de", "#91d95b", "#b8805f", "#cbbeeb"),
  legend_click = NULL,
  info
)
}
\arguments{
\item{aes}{list of selected adverse events}

\item{colors}{color vector for adverse events}
}
\description{
Drawing legends for adverse event colours in app
}
\keyword{internal}
