% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_and_impute_required_variables_missing_values.R
\name{calculate_and_impute_required_variables_missing_values}
\alias{calculate_and_impute_required_variables_missing_values}
\title{Impute missing values in adverse event data and return number of imputed data}
\usage{
calculate_and_impute_required_variables_missing_values(
  data,
  SUBJIDN = "SUBJIDN",
  LVDT = "LVDT",
  DTHDT = "DTHDT",
  TRTSDT = "TRTSDT",
  AEDECOD = "AEDECOD",
  AESTDY = "AESTDY",
  AETRTEMN = "AETRTEMN",
  AEENDY = "AEENDY",
  AESEVN = "AESEVN",
  severity_grading_flag
)
}
\arguments{
\item{data}{frame with adverse event data set}

\item{LVDT}{Last Visit Date as numeric (required)}

\item{DTHDT}{Death Date as numeric (required)}

\item{TRTSDT}{Treatment Start Date as numeric (required)}

\item{AEDECOD}{Adverse Event Code as character (required)}

\item{AESTDY}{Adverse Event Start Date as numeric (required)}

\item{AETRTEMN}{Adverse Event Treatment Emergency Flag as numeric (required)}

\item{AEENDY}{Adverse Event End date as numeric (required)}

\item{AESEVN}{Adverse Event Severity Grade Flag as numeric (required)}

\item{severity_grading_flag}{character value if 3 or 5 step grading is used}
}
\value{
data frame with imputed adverse event data
}
\description{
Impute missing values in adverse event data and return number of imputed data
}
