% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_number_ae_start_missing.R
\name{calculate_number_ae_start_missing}
\alias{calculate_number_ae_start_missing}
\title{Calculate the Number of Adverse Events Start Day Missing}
\usage{
calculate_number_ae_start_missing(dat, sel_aestdy, sel_aedecod, sel_aeendy)
}
\arguments{
\item{dat}{data frame with adverse event data}

\item{sel_aestdy}{character with adverse event start day variable name}

\item{sel_aedecod}{character with adverse event name}

\item{sel_aeendy}{character with adverse event end day variable name'}
}
\value{
numeric value with the number missing adverse event start days
}
\description{
Calculate the Number of Adverse Events Start Day Missing
}
