% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/fct_calculate_number_aes_not_treatment_emergent.R
\name{calculate_number_aes_not_treatment_emergent}
\alias{calculate_number_aes_not_treatment_emergent}
\title{Calculate the Number of Adverse Events Removed due to Missing Treatment Emergency Flag}
\usage{
calculate_number_aes_not_treatment_emergent(dat, sel_aedecod, sel_aetrtemn)
}
\arguments{
\item{dat}{data frame with adverse event data}

\item{sel_aedecod}{character with adverse event code variable name}

\item{sel_aetrtemn}{character with adverse event treatment emergency variable name}
}
\value{
numeric value with the number missing treatment emergency values
}
\description{
Calculate the Number of Adverse Events Removed due to Missing Treatment Emergency Flag
}
