% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_number_missing_last_visit_dates.R
\name{calculate_number_missing_last_visit_dates}
\alias{calculate_number_missing_last_visit_dates}
\title{Calculate the Number of Missing Last Visit Dates}
\usage{
calculate_number_missing_last_visit_dates(dat, sel_lvdt, sel_subjidn)
}
\arguments{
\item{dat}{data frame with adverse event data}

\item{sel_lvdt}{character with variable name of last visit date in data}
}
\value{
numeric value with the number missing last visit dates
}
\description{
Calculate the Number of Missing Last Visit Dates
}
