% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_number_missing_severity_flag.R
\name{calculate_number_missing_severity_flag}
\alias{calculate_number_missing_severity_flag}
\title{Calculate the Number of Missing Severity Grades}
\usage{
calculate_number_missing_severity_flag(
  dat,
  sel_aesevn = NULL,
  sel_aedecod,
  severity_grading_flag
)
}
\arguments{
\item{dat}{data frame with adverse event data}

\item{sel_aesevn}{character with severity grade variable name}

\item{sel_aedecod}{character with adverse event code variable name}

\item{severity_grading_flag}{logical value if 3 or 5 step grading is used}
}
\value{
numeric value with the number missing grades
}
\description{
Calculate the Number of Missing Severity Grades
}
