% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_and_prepare_patient_data.R
\name{check_colname_present}
\alias{check_colname_present}
\title{Checks if column name exists, throw an error otherwise}
\usage{
check_colname_present(colname, data)
}
\arguments{
\item{colname}{A character, the column name to check}

\item{data}{A data object}
}
\description{
Checks if column name exists, throw an error otherwise
}
