% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_data.R
\name{check_data}
\alias{check_data}
\title{check_data - Consistency checks on the data}
\usage{
check_data(ae_data, patients)
}
\arguments{
\item{ae_data}{adverse event dataset}

\item{patients}{patient dataset}
}
\description{
Runs consistency checks on the adverse event and patient data
}
\keyword{internal}
