% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_legend.R
\name{circle_legend}
\alias{circle_legend}
\title{circle_legend - Function to create legend for displayed subjects}
\usage{
circle_legend()
}
\description{
Creates legend object for base R plots
}
\keyword{internal}
