% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_event.R
\name{count_event}
\alias{count_event}
\title{count_event  Count the event number for every AE at spefified timepoint}
\usage{
count_event(total = tot, day = 1)
}
\arguments{
\item{total}{data set with ae data and patient data merged}

\item{day}{The study day of interest}
}
\description{
Count the event number for every AE at spefified timepoint
}
\keyword{internal}
