% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_prepare_patient_data.R
\name{filter_and_prepare_patient_data}
\alias{filter_and_prepare_patient_data}
\title{filter_and_prepare_patient_data - filter for safety flag and create required variables}
\usage{
filter_and_prepare_patient_data(
  data,
  SAFFN = SAFFN,
  LVDT = LVDT,
  TRTSDT = TRTSDT,
  TRT01A = TRT01A,
  SUBJIDN = SUBJIDN,
  DTHDT = DTHDT
)
}
\arguments{
\item{data}{adae (optional merged adsl) data set}

\item{SAFFN}{Safety Flag as numeric (required)}

\item{LVDT}{Last Visit Date as numeric (required)}

\item{TRTSDT}{Treatment Start Date as numeric (required)}

\item{TRT01A}{Treatment Group as character (required)}

\item{SUBJIDN}{Subject Id as numeric (required)}

\item{DTHDT}{Death Date as numeric (required)}
}
\description{
Creates a list with two data sets 'pat_data' and 'ae_data' in format which is used in the AdEPro Application
}
\keyword{internal}
