% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_safety_flag.R
\name{filter_for_safety_flag}
\alias{filter_for_safety_flag}
\title{Filter for Safety Flag and Create Required Variables}
\usage{
filter_for_safety_flag(dat, SAFFN)
}
\arguments{
\item{SAFFN}{Safety Flag as numeric (required)}

\item{data}{frame with adverse event data set}
}
\value{
data frame with filtered adverse event data
}
\description{
Filter for Safety Flag and Create Required Variables
}
