% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_and_prepare_patient_data.R
\name{filter_prepare_patient_data}
\alias{filter_prepare_patient_data}
\title{Filter for Safety Flag and Create Required Variables}
\usage{
filter_prepare_patient_data(data, saffn, lvdt, trtsdt, trt01a, subjidn, dthdt)
}
\arguments{
\item{data}{frame with adverse event data set}

\item{saffn}{Safety Flag as numeric (required)}

\item{lvdt}{Last Visit Date as numeric (required)}

\item{trtsdt}{Treatment Start Date as numeric (required)}

\item{trt01a}{Treatment Group as character (required)}

\item{subjidn}{Subject Id as numeric (required)}

\item{dthdt}{Death Date as numeric (required)}
}
\value{
data frame with filtered adverse event data
}
\description{
Filter for Safety Flag and Create Required Variables
}
