% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_number_of_unknown_aes.R
\name{get_number_of_unknown_aes}
\alias{get_number_of_unknown_aes}
\title{Get number of unknown AEs}
\usage{
get_number_of_unknown_aes(data, adae_data, adsl_data, sel_aedecod)
}
\arguments{
\item{data}{A dataframe}

\item{adae_data}{A dataframe. ADAE data}

\item{adsl_data}{A dataframe. ADSL data}

\item{sel_aedecod}{Character vector with names of selected AEs (AEDECOD)}
}
\description{
Get number of unknown AEs
}
