% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_initQ.R
\name{init_q}
\alias{init_q}
\title{init_q - generates classification matrix}
\usage{
init_q(ae_data)
}
\arguments{
\item{ae_data}{adverse event dataset}
}
\description{
Creates the classification matrix of adverse events such as treatment-emergent
}
\keyword{internal}
