% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_initQ.R
\name{initQ}
\alias{initQ}
\title{initQ - generates classification matrix}
\usage{
initQ(ae_data)
}
\arguments{
\item{ae_data}{adverse event dataset}
}
\description{
Creates the classification matrix of adverse events such as treatment-emergent
}
\keyword{internal}
