% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filter_and_prepare_patient_data.R
\name{is_convertible_to_date}
\alias{is_convertible_to_date}
\title{Checks if variable can be converted to date}
\usage{
is_convertible_to_date(variable)
}
\arguments{
\item{variable}{A vector, the variable}
}
\description{
Returns TRUE if at least a value can be parsed to date with anytime::anydate
}
