% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_order_patient.R
\name{order_patient}
\alias{order_patient}
\title{order_patient - Arranges patients by sequencing technique}
\usage{
order_patient(
  ae_data = ae_data,
  patients = patient_d,
  variables = input$varSeq,
  method_dist = "euclidean",
  method_seriate = input$methSeq
)
}
\arguments{
\item{ae_data}{Adverse event dataset}

\item{patients}{Patient dataset}

\item{variables}{Vector with Variable Names used to seriate}

\item{method_dist}{Character string with the name of the method to calculate the distance matrix}

\item{method_seriate}{Character string with the name of the method to seriate}
}
\description{
Arranges patients by sequencing technique (R Package required: 'dplyr', 'reshape2, 'seriation')
}
\keyword{internal}
