% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_legend2.R
\name{pie_legend2}
\alias{pie_legend2}
\title{pie_legend2 - Creates Legend for adverse events slices in AdEPro (version 2)}
\usage{
pie_legend2(
  tmp,
  aes,
  colors = c("#e43157", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33", "#a65628",
    "#f781bf", "#21d4de", "#91d95b", "#b8805f", "#cbbeeb"),
  legend_click = NULL,
  info,
  color_theme = TRUE
)
}
\arguments{
\item{aes}{list of selected adverse events}

\item{colors}{color vector for adverse events}

\item{color_theme}{logical value for dark/light theme}
}
\description{
Drawing legends for adverse event symbols in app
}
\keyword{internal}
