% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data - read SAS or CSV raw data and prepare the data sets}
\usage{
prepare_data(
  dat,
  SUBJIDN = "SUBJIDN",
  TRT01A = "TRT01A",
  SAFFN = "SAFFN",
  LVDT = "LVDT",
  DTHDT = "DTHDT",
  TRTSDT = "TRTSDT",
  AEDECOD = "AEDECOD",
  AESTDY = "AESTDY",
  AETRTEMN = "AETRTEMN",
  AEENDY = "AEENDY",
  AESEVN = "AESEVN",
  AESERN = "AESERN",
  AERELN = "AERELN",
  AERELPRN = "AERELPRN",
  AEACNN = "AEACNN",
  adsl_data = NULL
)
}
\arguments{
\item{dat}{adae data set.}

\item{SUBJIDN}{Subject Id as numeric (required)}

\item{TRT01A}{Treatment Group as character (required)}

\item{SAFFN}{Safety Flag as numeric (required)}

\item{LVDT}{Last Visit Date as numeric (required)}

\item{DTHDT}{Death Date as numeric (required)}

\item{TRTSDT}{Treatment Start Date as numeric (required)}

\item{AEDECOD}{Adverse Event Code as character (required)}

\item{AESTDY}{Adverse Event Start Date as numeric (required)}

\item{AETRTEMN}{Adverse Event Treatment Emergency Flag as numeric (required)}

\item{AEENDY}{Adverse Event End date as numeric (required)}

\item{AESEVN}{Adverse Event Severity Grade Flag as numeric (required)}

\item{AESERN}{Adverse Event Serious Flag as numeric (optional)}

\item{AERELN}{Adverse Event Related Flag as numeric (optional)}

\item{AERELPRN}{Adverse Event      Flag as numeric (optional)}

\item{AEACNN}{Adverse Event        Flag as numeric (optional)}

\item{adsl_data}{Subject Level data set (optional)}
}
\description{
Creates a list with two data sets 'pat_data' and 'ae_data' in format which is used in the AdEPro Application
}
\keyword{internal}
