% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data_for_adepro.R
\name{prepare_data_for_adepro}
\alias{prepare_data_for_adepro}
\title{prepare_data - read SAS or CSV raw data and prepare the data sets}
\usage{
prepare_data_for_adepro(
  dat,
  subjidn = "SUBJIDN",
  trt01a = "TRT01A",
  saffn = "SAFFN",
  lvdt = "LVDT",
  dthdt = "DTHDT",
  trtsdt = "TRTSDT",
  aedecod = "AEDECOD",
  aestdy = "AESTDY",
  aetrtemn = "AETRTEMN",
  aeendy = "AEENDY",
  aesevn = "AESEVN",
  aesern = "AESERN",
  aereln = "AERELN",
  aerelprn = "AERELPRN",
  aeacnn = "AEACNN"
)
}
\arguments{
\item{dat}{adae data set.}

\item{subjidn}{Subject Id as numeric (required)}

\item{trt01a}{Treatment Group as character (required)}

\item{saffn}{Safety Flag as numeric (required)}

\item{lvdt}{Last Visit Date as numeric (required)}

\item{dthdt}{Death Date as numeric (required)}

\item{trtsdt}{Treatment Start Date as numeric (required)}

\item{aedecod}{Adverse Event Code as character (required)}

\item{aestdy}{Adverse Event Start Date as numeric (required)}

\item{aetrtemn}{Adverse Event Treatment Emergency Flag as numeric (required)}

\item{aeendy}{Adverse Event End date as numeric (required)}

\item{aesevn}{Adverse Event Severity Grade Flag as numeric (required)}

\item{aesern}{Adverse Event Serious Flag as numeric (optional)}

\item{aereln}{Adverse Event Related Flag as numeric (optional)}

\item{aerelprn}{Adverse Event Flag as numeric (optional)}

\item{aeacnn}{Adverse Event Flag as numeric (optional)}
}
\description{
Creates a list with two data sets 'pat_data' and 'ae_data' in format which is
used in the AdEPro Application
}
\keyword{internal}
