% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_preproc_patients.R
\name{preproc_patients}
\alias{preproc_patients}
\title{preproc_patients - Preprocessing Patient Data}
\usage{
preproc_patients(patients, height)
}
\arguments{
\item{patients}{patient dataset}

\item{height}{number of circles to be displayed on the vertical axis}
}
\description{
Preprocesses patient dataset
}
\keyword{internal}
