% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server part of the AdEPro application}
\usage{
server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\value{
No return value. Server part of the app, used in launch_adepro-function.
}
\description{
Server part of the AdEPro application
}
\keyword{internal}
