% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_global_params.R
\name{set_global_params}
\alias{set_global_params}
\title{set_global_params - sets all global parameters}
\usage{
set_global_params(
  ae_data,
  patients,
  title = NULL,
  height = NULL,
  treatment = NULL
)
}
\arguments{
\item{ae_data}{adverse event dataset}

\item{patients}{patient dataset}

\item{title}{titles for the treatment groups}

\item{height}{number of circles on the vertical axis}
}
\description{
Function that sets all global parameters
}
\keyword{internal}
