% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_set_group_lines.R
\name{set_group_lines}
\alias{set_group_lines}
\title{set_group_lines - determines coordinates for the separating lines}
\usage{
set_group_lines(patients, height, treatment)
}
\arguments{
\item{patients}{patient dataset}

\item{height}{number of circles on the vertical axis}
}
\description{
Calculates coordinates for separating lines between the treatment groups
}
\keyword{internal}
