% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vector_layout.R
\name{set_vector_layout}
\alias{set_vector_layout}
\title{set_vector_layout - creates vector layout}
\usage{
set_vector_layout(patients, height)
}
\arguments{
\item{patients}{patient dataset}

\item{height}{number of circles on the vertical axis}
}
\description{
creates vector layout for circles considering number of rows and treatment groups
}
\keyword{internal}
