% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_set_width.R
\name{set_width}
\alias{set_width}
\title{set_width - calculates width (number of columns with circles)}
\usage{
set_width(patients, height)
}
\arguments{
\item{patients}{patient dataset}

\item{height}{number of circles on the vertical axis}
}
\description{
Calculates vector with 1's with its length as the number of circles on the
vertical axis
}
\keyword{internal}
