% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone.R
\name{tone}
\alias{tone}
\title{tone - Function that plays two different sounds for AEs}
\usage{
tone(no1 = 0, no2 = 0, d = 0.7)
}
\arguments{
\item{no1}{classification of the frequency of the AE in the first treatment group
(0=no AE chosen, 1=no occurrence, 2=rare, 3=occasionally, 4=often)}

\item{no2}{classification of the frequency of the AE in the second treatment group (
0=no AE chosen, 1=no occurrence, 2=rare, 3=occasionally, 4=often)}

\item{d}{Duration of the sounds (numeric, in seconds)}
}
\description{
Plays two different sounds (after one another) for treatment group comparison for a specific AE
}
\keyword{internal}
