% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{User Interface of the AdEPro application}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
No return value. User interface part of the app, used in launch_adepro-function.
}
\description{
User Interface of the AdEPro application
}
\keyword{internal}
