% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adepro_slice_plot.R
\name{adepro_slice_plot}
\alias{adepro_slice_plot}
\title{adepro_slice_plot - function to create pie chart graph}
\usage{
adepro_slice_plot(
  data,
  patients,
  ae_list,
  global_params,
  height,
  width,
  xlines,
  ylines,
  xval,
  title,
  subgroup,
  slider,
  subjidn,
  adepro_colors = c("#e43157", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33",
    "#a65628", "#f781bf", "#21d4de", "#91d95b", "#b8805f", "#cbbeeb")
)
}
\arguments{
\item{data}{adverse event data set}

\item{patients}{patient data set}

\item{ae_list}{list with selected adverse events}

\item{global_params}{global parameter}

\item{height}{height size}

\item{width}{width size}

\item{xlines}{lines position x}

\item{ylines}{lines position y}

\item{xval}{label x-axis}

\item{title}{graph title}

\item{slider}{day}

\item{subjidn}{subjidn variable}

\item{adepro_colors}{colors used in adepro (max 12)}

\item{subroup}{selected subgroup}
}
\description{
Create pie chart graph for app adepro
}
\keyword{internal}
