% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_number_ae_end_missing.R
\name{calculate_number_ae_end_missing}
\alias{calculate_number_ae_end_missing}
\title{Calculate the Number of Adverse Events End Day Missing}
\usage{
calculate_number_ae_end_missing(dat, sel_aeendy, sel_aedecod, sel_aestdy)
}
\arguments{
\item{dat}{data frame with adverse event data}

\item{sel_aeendy}{character with adverse event end day variable name}

\item{sel_aedecod}{character with adverse event code variable name}

\item{sel_aestdy}{character with adverse event start day variable name}
}
\value{
numeric value with the number missing adverse event end days
}
\description{
Calculate the Number of Adverse Events End Day Missing
}
