% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_number_unknown_aes.R
\name{calculate_number_unknown_aes}
\alias{calculate_number_unknown_aes}
\title{Calculate the Number of Unknown Adverse Events}
\usage{
calculate_number_unknown_aes(dat, sel_aedecod, sel_aestdy, sel_aeendy)
}
\arguments{
\item{dat}{data frame with adverse event data}

\item{sel_aedecod}{character with adverse event code variable name}

\item{sel_aestdy}{character with adverse event start day variable name}

\item{sel_aeendy}{character with adverse event end day variable name}
}
\value{
numeric value with the number empty adverse events ("")
}
\description{
Calculate the Number of Unknown Adverse Events
}
