% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_join_adae_and_adsl.R
\name{join_adae_and_adsl}
\alias{join_adae_and_adsl}
\title{Join Adverse Event Data and Subject Level Data Sets}
\usage{
join_adae_and_adsl(dat_adae, dat_adsl, SUBJIDN)
}
\arguments{
\item{dat_adae}{data frame with adverse event data set}

\item{dat_adsl}{data frame with subject level data set}

\item{SUBJIDN}{character with subject identifier variable (required column in both data sets)}
}
\value{
merged data set with adae and adsl information
}
\description{
Join Adverse Event Data and Subject Level Data Sets
}
