% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_ae.R
\name{preproc_ae}
\alias{preproc_ae}
\title{preproc_ae - Preprocessing adverse event data}
\usage{
preproc_ae(ae_data)
}
\arguments{
\item{ae_data}{adverse event dataset}
}
\description{
Preprocesses adverse event dataset
}
\keyword{internal}
